01-MCP-Files-Testing/01-starter/pyproject.toml
[project]
name = "solution"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.19.0",
    "pillow>=12.0.0",
]